Componentizacion de los botones del Dashboard de Usuario V1.1
