+++ src/Components/Shared/DashboardOption/dashboard-option.module.css
@@ -1,0 +1,33 @@
+.optionCard {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    width: 120px;
+    height: 120px;
+    background-color: #f4f4f4;
+    border-radius: 10px;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+    transition: transform 0.2s ease, box-shadow 0.2s ease;
+    cursor: pointer;
+    text-align: center;
+    padding: 10px;
+  }
+  
+  .optionCard:hover {
+    transform: scale(1.05);
+    box-shadow: 0 6px 10px rgba(0, 0, 0, 0.15);
+  }
+  
+  .icon {
+    font-size: 2rem;
+    margin-bottom: 8px;
+    color: #555;
+  }
+  
+  p {
+    font-size: 1rem;
+    font-weight: bold;
+    color: #333;
+  }
+
